Small changes to RestProxy and AzureProxy
