+++ azure-client-runtime/src/main/java/com/microsoft/azure/v2/AzureServiceClient.java
@@ -25,6 +25,6 @@ public abstract class AzureServiceClient extends ServiceClient {
 
     @Override
     protected SerializerAdapter<?> createSerializerAdapter() {
-        return AzureProxy.defaultSerializer();
+        return AzureProxy.createDefaultSerializer();
     }
 }
