+++ client-runtime/src/main/java/com/microsoft/rest/v2/RestProxy.java
@@ -6,7 +6,6 @@
 
 package com.microsoft.rest.v2;
 
-import com.google.common.hash.Hashing;
 import com.google.common.reflect.TypeToken;
 import com.microsoft.rest.v2.credentials.ServiceClientCredentials;
 import com.microsoft.rest.v2.http.ContentType;
@@ -20,11 +19,9 @@ import com.microsoft.rest.v2.http.HttpRequest;
 import com.microsoft.rest.v2.http.HttpResponse;
 import com.microsoft.rest.v2.http.NettyClient;
 import com.microsoft.rest.v2.http.UrlBuilder;
-import com.microsoft.rest.v2.policy.AddCookiesPolicy;
 import com.microsoft.rest.v2.policy.CredentialsPolicy;
 import com.microsoft.rest.v2.policy.LoggingPolicy;
 import com.microsoft.rest.v2.policy.RequestPolicy;
-import com.microsoft.rest.v2.policy.RetryPolicy;
 import com.microsoft.rest.v2.protocol.SerializerAdapter;
 import com.microsoft.rest.v2.protocol.SerializerAdapter.Encoding;
 import com.microsoft.rest.v2.protocol.TypeFactory;
@@ -45,8 +42,6 @@ import java.lang.reflect.Method;
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Proxy;
 import java.lang.reflect.Type;
-import java.net.NetworkInterface;
-import java.util.Enumeration;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -542,110 +537,52 @@ public class RestProxy implements InvocationHandler {
         return getTypeArguments(type)[0];
     }
 
-    private static <T> String getDefaultUserAgentString(Class<?> swaggerInterface) {
-        final String packageImplementationVersion = swaggerInterface.getPackage().getImplementationVersion();
-        final String operatingSystem = operatingSystem();
-        final String macAddressHash = macAddressHash();
-        final String javaVersion = javaVersion();
-        return String.format("Azure-SDK-For-Java/%s OS:%s MacAddressHash:%s Java:%s",
-                packageImplementationVersion,
-                operatingSystem,
-                macAddressHash,
-                javaVersion);
-    }
-
-    private static String operatingSystem;
-    private static String operatingSystem() {
-        if (operatingSystem == null) {
-            operatingSystem = System.getProperty("os.name") + "/" + System.getProperty("os.version");
-        }
-        return operatingSystem;
-    }
-
-    private static String macAddressHash;
-    private static String macAddressHash() {
-        if (macAddressHash == null) {
-            byte[] macBytes = null;
-            try {
-                Enumeration<NetworkInterface> networks = NetworkInterface.getNetworkInterfaces();
-                while (networks.hasMoreElements()) {
-                    NetworkInterface network = networks.nextElement();
-                    macBytes = network.getHardwareAddress();
-
-                    if (macBytes != null) {
-
-                        break;
-                    }
-                }
-            } catch (Throwable t) {
-                // It's okay ignore mac address hash telemetry
-            }
-
-            if (macBytes == null) {
-                macBytes = "Unknown".getBytes();
-            }
-
-            macAddressHash = Hashing.sha256().hashBytes(macBytes).toString();
-        }
-        return macAddressHash;
-    }
-
-    private static String javaVersion;
-    private static String javaVersion() {
-        if (javaVersion == null) {
-            final String versionProperty = System.getProperty("java.version");
-            javaVersion = versionProperty != null ? versionProperty : "Unknown";
-        }
-        return javaVersion;
-    }
-
     /**
-     * Get the default serializer.
+     * Create an instance of the default serializer.
      * @return the default serializer.
      */
-    public static SerializerAdapter<?> defaultSerializer() {
+    public static SerializerAdapter<?> createDefaultSerializer() {
         return new JacksonAdapter();
     }
 
-    /**
-     * Create the default HttpPipeline.
-     * @param swaggerInterface The interface that the pipeline will use to generate a user-agent
-     *                         string.
-     * @return the default HttpPipeline.
-     */
-    public static HttpPipeline defaultPipeline(Class<?> swaggerInterface) {
-        return defaultPipeline(swaggerInterface, (RequestPolicy.Factory) null);
+    private static HttpClient.Factory defaultHttpClientFactory;
+    private static HttpClient.Factory defaultHttpClientFactory() {
+        if (defaultHttpClientFactory == null) {
+            defaultHttpClientFactory = new NettyClient.Factory();
+        }
+        return defaultHttpClientFactory;
     }
 
     /**
      * Create the default HttpPipeline.
-     * @param swaggerInterface The interface that the pipeline will use to generate a user-agent
-     *                         string.
+     * @return the default HttpPipeline.
+     */
+    public static HttpPipeline createDefaultPipeline() {
+        return createDefaultPipeline((RequestPolicy.Factory) null);
+    }
+
+    /**
+     * Create the default HttpPipeline.
      * @param credentials The credentials to use to apply authentication to the pipeline.
      * @return the default HttpPipeline.
      */
-    public static HttpPipeline defaultPipeline(Class<?> swaggerInterface, ServiceClientCredentials credentials) {
-        return defaultPipeline(swaggerInterface, new CredentialsPolicy.Factory(credentials));
+    public static HttpPipeline createDefaultPipeline(ServiceClientCredentials credentials) {
+        return createDefaultPipeline(new CredentialsPolicy.Factory(credentials));
     }
 
     /**
      * Create the default HttpPipeline.
-     * @param swaggerInterface The interface that the pipeline will use to generate a user-agent
-     *                         string.
      * @param credentialsPolicy The credentials policy factory to use to apply authentication to the
      *                          pipeline.
      * @return the default HttpPipeline.
      */
-    public static HttpPipeline defaultPipeline(Class<?> swaggerInterface, RequestPolicy.Factory credentialsPolicy) {
-        final HttpClient httpClient = new NettyClient.Factory().create(null);
+    public static HttpPipeline createDefaultPipeline(RequestPolicy.Factory credentialsPolicy) {
+        final HttpClient httpClient = defaultHttpClientFactory().create(null);
         final HttpPipeline.Builder builder = new HttpPipeline.Builder(httpClient);
-        builder.withUserAgent(getDefaultUserAgentString(swaggerInterface));
-        builder.withRequestPolicy(new RetryPolicy.Factory());
-        builder.withRequestPolicy(new AddCookiesPolicy.Factory());
         if (credentialsPolicy != null) {
             builder.withRequestPolicy(credentialsPolicy);
         }
-        builder.withRequestPolicy(new LoggingPolicy.Factory(LogLevel.HEADERS));
+        builder.withRequestPolicy(new LoggingPolicy.Factory(LogLevel.BODY_AND_HEADERS));
         return builder.build();
     }
 
@@ -657,7 +594,7 @@ public class RestProxy implements InvocationHandler {
      */
     @SuppressWarnings("unchecked")
     public static <A> A create(Class<A> swaggerInterface) {
-        return create(swaggerInterface, defaultPipeline(swaggerInterface), defaultSerializer());
+        return create(swaggerInterface, createDefaultPipeline(), createDefaultSerializer());
     }
 
     /**
@@ -670,7 +607,7 @@ public class RestProxy implements InvocationHandler {
      */
     @SuppressWarnings("unchecked")
     public static <A> A create(Class<A> swaggerInterface, HttpPipeline httpPipeline) {
-        return create(swaggerInterface, httpPipeline, defaultSerializer());
+        return create(swaggerInterface, httpPipeline, createDefaultSerializer());
     }
 
     /**
